Keep track of errors in individual addresses under round-robin DNS so they are not re-used, except that if all proxies are used up then the errors are erased for the next try.  This was already being done for explicitly listed proxies & servers when doing load balancing.
When doing load balancing, which randomly selects between the listed
proxies or servers, don't give preferential treatment to those listed
after the failed proxies or servers. The random process used to
include the failed addresses and select the one listed after them, so
for example if there was one failure out of three proxies, one would
get 2/3 of the connections and the other 1/3.