Support sending Last-Modified header and accepting If-Modified-Since. Squid automatically sends the latter header when the cached object expires if the former header exists.  If the timestamp for a query (which is per-table) hasn't changed, return http code 304 Not Modified.  This is very tricky because if the timestamp for the table isn't cached, need to acquire the database and that can take a long time and can require sending keepalive messages which in turn requires already sending out the header.  In the worst case the answer is sent without the Last-Modified header, so the object will just expire and squid will try again later with a regular (non If-Modified-Since) request.
Switch to using Cache-Control: max-age instead of Expires header
because it is simpler to calculate and manage.   Make the handling of
global error conditions more robust.  Also make the errors logged for
exeptions be more concise by using trim() on them and by removing a
leading "java.lang." from the beginning if it's there.