Add support for setting a "max_age" at the end of the payload if the header has already been sent with a longer Cache-control: max-age than is determined later.  Be more consistent in applying the shorter cache time too: previously the shorter time was indicated by signalling a global error, but the code was duplicated twice and not applied in third situation; now it is done in a function error_payload_end.
Also move the setting of the Last-Modified header earlier when it was
internally cached but the data still needed to be read; formerly
Last-Modified was not set until until after the database was acquired,
which was too late if keepalive messages were sent.  It used to be
delayed because of the potential for errors, but really there was no
need for the delay since a global error would clean up such a badly
cached error message anyway.  Formerly global error cleanups did have
a potential for server overload, however, so it was good to avoid them
when possible, but now with the new "max_age" keyword and new clients
delaying their hard refreshes that overload potential is removed.

Also set max_age (in either the header or end of the payload as needed)
to the shorter "error" time when a query returns empty results.