Add support for forever ttl.
Add support for max_age in response, and base it on the Age http header.
If it is a global error without max_age, assume the max_age is 5 minutes.
This is much safer than using 0 seconds as it used to because many
clients could receive the error around the same time and overload the
server.

Use ctime_r instead of ctime for thread safety.